remove 0xff bytes for unused space
Removing the hard-coded 0xff allows developers to add code without
worrying about overflowing the byte section.  This is because the
linker will fill any unused space with 0xff.  As part of this change
the interrupt segment has been moved to more closely match where it
ends up in the rom.